+++ packages/nrf_ble_dfu_ui/lib/nrf_ble_dfu_ui.dart
@@ -6,6 +6,7 @@ import 'src/dfu_ui_manager.dart';
 export 'package:nrf_ble_dfu/nrf_ble_dfu.dart';
 export 'src/widget/widget.dart';
 export 'src/dfu_ui_manager.dart';
+export 'src/auto_dfu_controller.dart';
 
 Future<void> initNrfBleDfuForFlutter() async {
   // 1. Initialize the Core package